Fix a few warnings due to recent changes in roundToZero().
git-svn-id: svn://svn.berlios.de/gale/trunk@745 2d798ef8-6e06-0410-941a-9092aa770104

+++ gale2/include/gale/math/color.h
@@ -122,7 +122,7 @@ struct ColorModel
         float R,G,B;
 
         if (S>0) {
-            long i=roundToZero(H);
+            long long i=roundToZero(H);
             float f=H-i;
 
             float p=V*(1-S);
