+++ gale2/source/mesh_factory.cpp
@@ -221,7 +221,7 @@ void Mesh::Factory::populateNeighborhood(Mesh* mesh,float distance,int valence)
 
                 // Get the oriented angle between the neighbors in the plane.
                 double oa=up.orientedAngle(vp,r)*Constd::RAD_TO_DEG();
-                n=roundToEven(oa/(360.0/valence));
+                n=static_cast<int>(roundToEven(oa/(360.0/valence)));
             }
 
             // Store the neighbor's vertex index at the calculated position.
